Refactoring and moving classes to /nightly
