+++ Jenkinsfile_nightly
@@ -50,8 +50,8 @@ withNightlyPipeline(type, product, component) {
     println("### Set test_environment.. ###")
     env.test_environment = 'aat'
     after('DependencyCheckNightly') {
-        stage('PriorityFunctional') {
-            println("~~~~~~~~~~~~~~~~~  Starting Priority Functional :aat ~~~~~~~~~~")
+        stage('LowerPriorityFunctional') {
+            println("~~~~~~~~~~~~~~~~~  Starting Lower Priority Functional :aat ~~~~~~~~~~")
             sh("./gradlew --no-daemon --init-script init.gradle priorityFunctional")
             steps.archiveArtifacts allowEmptyArchive: true, artifacts: '**/site/serenity/**/*'
             steps.archiveArtifacts allowEmptyArchive: true, artifacts: 'build/reports/**/*'
@@ -59,7 +59,7 @@ withNightlyPipeline(type, product, component) {
         }
     }
      enableFullFunctionalTest()
-    // enableMutationTest()
-    // enableSecurityScan()
+     //enableMutationTest()
+     //enableSecurityScan()
 }
