+++ build.gradle
@@ -399,9 +399,7 @@ task functional(type: Test, description: 'Runs the functional tests.', group: 'V
     maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
     forkEvery = 10
 
-    exclude "uk/gov/hmcts/reform/divorce/maintenance/**"
-    exclude "uk/gov/hmcts/reform/divorce/draft/**"
-
+    exclude "uk/gov/hmcts/reform/divorce/maintenance/nightly/**"
 
     setTestClassesDirs(sourceSets.integrationTest.output.classesDirs)
     setClasspath(sourceSets.integrationTest.runtimeClasspath)
@@ -422,9 +420,8 @@ task priorityFunctional(type: Test, description: 'Runs the Priority Functional t
     maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
     forkEvery = 10
 
-    include "uk/gov/hmcts/reform/divorce/maintenance/**"
-    include "uk/gov/hmcts/reform/divorce/draft/**"
-
+    include "uk/gov/hmcts/reform/divorce/maintenance/nightly/**"
+    
     setTestClassesDirs(sourceSets.integrationTest.output.classesDirs)
     setClasspath(sourceSets.integrationTest.runtimeClasspath)
