--- src/integrationTest/java/uk/gov/hmcts/reform/divorce/maintenance/MovedProcessPbaPaymentTest.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.reform.divorce.maintenance;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-public class MovedProcessPbaPaymentTest {
-
-    @Test
-    public void dummy() {
-        Assert.assertTrue(true);
-    }
-
-}
