+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/maintenance/nightly/LinkRespondentTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.maintenance;
+package uk.gov.hmcts.reform.divorce.maintenance.nightly;
 
 import io.restassured.response.Response;
 import org.apache.commons.lang3.tuple.ImmutablePair;
