+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/maintenance/nightly/MakeCaseEligibleForDATest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.maintenance;
+package uk.gov.hmcts.reform.divorce.maintenance.nightly;
 
 import io.restassured.response.Response;
 import lombok.extern.slf4j.Slf4j;
