+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/maintenance/nightly/SubmitDnCaseTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.maintenance;
+package uk.gov.hmcts.reform.divorce.maintenance.nightly;
 
 import io.restassured.response.Response;
 import org.junit.Before;
