+++ src/integrationTest/resources/fixtures/da-granted/da-granted-offline.json
@@ -25,9 +25,9 @@
           "id": "6233e5eb-15d7-4f35-8576-d9713a343ab5",
           "value": {
             "DocumentLink": {
-              "document_url": "http://dm-store-aat.service.core-compute-aat.internal/documents/812f2709-7891-4f7e-835e-11a84a1fa008",
+              "document_url": "http://dm-store-aat.service.core-compute-aat.internal/documents/${documentId}",
               "document_filename": "decreeAbsolute1590653987162971.pdf",
-              "document_binary_url": "http://dm-store-aat.service.core-compute-aat.internal/documents/812f2709-7891-4f7e-835e-11a84a1fa008/binary"
+              "document_binary_url": "http://dm-store-aat.service.core-compute-aat.internal/documents/${documentId}/binary"
             },
             "DocumentType": "daGranted",
             "DocumentComment": null,
